+++ e2e/e2etests/test_bitcoin_withdraw_rbf.go
@@ -34,7 +34,9 @@ func TestBitcoinWithdrawRBF(r *runner.E2ERunner, args []string) {
 	// initiate a withdraw CCTX, and wait for CCTX creation
 	tx := r.WithdrawBTC(to, amount, gatewayzevm.RevertOptions{OnRevertGasLimit: big.NewInt(0)}, true)
 	utils.WaitForZetaBlocks(r.Ctx, r, r.ZEVMClient, 1, 20*time.Second)
-	cctx := utils.GetCCTXByInboundHash(r.Ctx, r.CctxClient, tx.Hash().Hex())
+	cctxs := utils.GetCCTXByInboundHash(r.Ctx, r.CctxClient, tx.Hash().Hex())
+	require.Len(r, cctxs, 1)
+	cctx := &cctxs[0]
 
 	// wait for the 1st outbound tracker hash to come in
 	nonce := cctx.GetCurrentOutboundParam().TssNonce